Реализована авторизация в черновиках

+++ src/main/java/it/college/AnnouncementBackend/routes/drafts/servicies/DraftService.java
@@ -56,8 +56,19 @@ public class DraftService {
 
     public ResponseEntity findDraftByUUID(String token, String uuid){
         try {
-            Optional<Announcement> announcement = this.repository.findById(java.util.UUID.fromString(uuid));
-            return new ResponseEntity(announcement.get(), HttpStatus.OK);
+            Optional<Announcement> OptionalAnnouncement = this.repository.findById(java.util.UUID.fromString(uuid));
+
+            if (OptionalAnnouncement.isEmpty()){
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            }
+
+            Announcement announcement = OptionalAnnouncement.get();
+
+           if (!announcement.getAuthorID().equals(auth.auth(token))) {
+               return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+           }
+
+            return new ResponseEntity(announcement, HttpStatus.OK);
         }catch (DataIntegrityViolationException e) {
             System.err.println("Data: " + e.getMessage());
             return ResponseEntity.status(HttpStatus.CONFLICT).build();
@@ -79,15 +90,21 @@ public class DraftService {
         try {
             Optional<Announcement> current = this.repository.findById(UUID.fromString(uuid));
 
+            String author = auth.auth(token);
+
             if (current.isEmpty()){
                 Announcement announcement = mapper.getMapper().map(payload, Announcement.class);
 
                 announcement.setStatus(AStatus.Draft);
                 announcement.setZoneDateTime(LocalDateTime.now());
-                announcement.setAuthorID(token);
+                announcement.setAuthorID(author);
 
                 this.repository.save(announcement);
                 return new ResponseEntity(announcement, HttpStatus.OK);
+            }
+
+            if (!current.get().getAuthorID().equals(author)){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
             }
 
             current.get().setJsonContent(payload.getJsonContent());
@@ -113,6 +130,12 @@ public class DraftService {
 
     public ResponseEntity deleteDraftByUUID(String token, String uuid){
         try {
+            Announcement announcement = this.repository.findById(UUID.fromString(uuid)).get();
+
+            if (!announcement.getAuthorID().equals(auth.auth(token))){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
             this.repository.deleteById(UUID.fromString(uuid));
             return ResponseEntity.ok().build();
         }catch (DataIntegrityViolationException e) {
